[#56005352]
Add #initialize to payload modules and non-payload classes with an
attributes={} argument so that they work with the calling convention in
the module loading in metasploit-framework.  The
Metasploit::Model::Module::Ancestor#contents composing was becoming hard
to following, so I converted from using f.puts and having to re-write
the file for Metasploit::Model::Module::Class, to having .rb.erb
templates that can just be re-run for Metasploit::Model::Module::Class.