Ensure Metasploit::Model::Translation is included and not extended
[#56005352]

Metasploit::Model::Reference was extending instead of including
Metasploit::Model::Translation as was the included block for
Metasploit::Model::Instance.  The 'Metasploit::Model::Translation'
shared example did not catch this bug (it was caught in
metasploit_data_model's specs), so strength the shared example to check
the dependencies of the metasploit_model_ancestor to ensure
Metasploit::Model::Translation is being included and it's not coming
from Metasploit::Model::Base (which is why the spec was passing).