Tab completion helpers
MSP-9007

Allow direct setting of
`Metasploit::Framework::Search::Query#formatted_operations` so that
shell splitting can occur in msfconsole.  Allow removal of all
operations on a given operator so that operation being tab completed
will not be restricted by previous operations using the same operator.
Without removing the operator, the tab completion would only return the
pre-existing value because the pluck wouldn't be properly unioned with
the pre-existing values as is the case when a full operation is used.