Fix union operations without values
MSP-8986

`Metasploit::Model::Search::Operation#parse` was passing the formatted
value as `nil` if the formatted operation did not contain a ':', but the
union operator can't handle that and should have to, so pass '' and
treat '<formatted_operator>' the same as '<formatted_operator>:'.  For
union operations, the children weren't being validated, so processing
was continuing when it shouldn't.