Add support for render locals
MSP-9228

In addition the locals passed to the template instance, each call to
 #render can also be passed additional locals, which is useful for
rendering elements of a collection.  Support per-#render locals also
changes how locals are accessed: instead of being looked up using
method_missing, the binding of #result is duplicated and then the
instance level #locals and the locals options passed to #render and
therefore #result, are all injected into the duplicated binding, which
is then passed to ERB#result.