Generalize groups for operators and operations
MSP-10637

Extract the general behavior of Operator::Union to
Operator::Group::Base.  Then have Operator::Group::Union and
Operator::Group::Base as subclasses so that visit can be tuned based the
class being a union or intersection.  Do the same changes for
Operations.  Operator and Operation level intersection are needed so
that an intersection can be forced when the operator is the same as the
normal query.tree lock does a group_by(:operator) that unions operations
with the same operator.